+++ Source/Kernel/Events.Projections/MongoDB/MongoDBProjectionPositions.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Cratis.Events;
 using Cratis.Events.Projections;
 using Cratis.Extensions.MongoDB;
 using MongoDB.Driver;