+++ Source/Kernel/Events.Schemas.Api/EventTypes.cs
@@ -1,0 +1,39 @@
+// Copyright (c) Cratis. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Text.Json;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Cratis.Events.Schemas.API
+{
+    [Route("/api/events/types")]
+    public class EventTypes : Controller
+    {
+        readonly ISchemaStore _schemaStore;
+
+        public EventTypes(ISchemaStore schemaStore)
+        {
+            _schemaStore = schemaStore;
+        }
+
+        [HttpGet]
+        public async Task<IEnumerable<EventType>> AllEventTypes()
+        {
+            var schemas = await _schemaStore.GetLatestForAllEventTypes();
+
+            return schemas.Select(_ =>
+                new EventType(
+                    _.Type.Id.ToString(),
+                    _.Schema.GetDisplayName(),
+                    _.Schema.GetGeneration()));
+        }
+
+        [HttpGet("schemas/{eventTypeId}")]
+        public async Task<IEnumerable<JsonDocument>> GenerationSchemasForType(
+            [FromRoute] string eventTypeId)
+        {
+            var schemas = await _schemaStore.GetAllGenerationsForEventType(new (eventTypeId, 1));
+            return schemas.Select(_ => JsonDocument.Parse(_.Schema.ToJson()));
+        }
+    }
+}