+++ Specifications/Kernel/Events.Projections/for_Projection/when_getting_key_resolver_for_event_type.cs
@@ -2,13 +2,13 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Cratis.Properties;
-using Newtonsoft.Json.Schema;
+using NJsonSchema;
 
 namespace Cratis.Events.Projections.for_Projection
 {
     public class when_getting_key_resolver_for_event_type : Specification
     {
-        static EventType event_type = "993888cc-a9c5-4d56-ae21-f732159feec7";
+        static EventType event_type = new("993888cc-a9c5-4d56-ae21-f732159feec7", 1);
         Projection projection;
         ValueProvider<Event> expected;
         ValueProvider<Event> result;
@@ -20,7 +20,7 @@ namespace Cratis.Events.Projections.for_Projection
                 "0b7325dd-7a25-4681-9ab7-c387a6073547",
                 string.Empty,
                 string.Empty,
-                new Model(string.Empty, new JSchema()),
+                new Model(string.Empty, new JsonSchema()),
                 new[] {
                     new EventTypeWithKeyResolver(event_type, expected)
                 },