+++ Specifications/Kernel/Events.Projections/for_ProjectionExtensions/when_applying_from_filter/and_issuing_event_with_type_that_is_not_included.cs
@@ -1,0 +1,20 @@
+// Copyright (c) Cratis. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Cratis.Events.Projections.for_ProjectionExtensions.when_applying_from_filter
+{
+    public class and_issuing_event_with_type_that_is_not_included : given.an_observable_and_event_setup
+    {
+        IObservable<EventContext> filtered;
+
+        void Establish()
+        {
+            filtered = observable.From(new EventType("745a8adb-aec5-4bf7-af29-b23d14e5c7bc", 1));
+            filtered.Subscribe(_ => received.Add(_));
+        }
+
+        void Because() => observable.OnNext(event_context);
+
+        [Fact] void should_not_get_any_events() => received.ShouldBeEmpty();
+    }
+}